+++ engines/stack.go
@@ -33,8 +33,8 @@ func FetchCode(query string, page int) ([]entities.Stack, error) {
 
 	for i := range stacks {
 		stacks[i].Title = html.UnescapeString(stacks[i].Title)
-		stacks[i].ScoreStr = utilities.FormatHumanReadable(uint64(stacks[i].Score))
-		stacks[i].ViewCountStr = utilities.FormatHumanReadable(stacks[i].ViewCount)
+		stacks[i].ScoreStr = utilities.FormatHumanReadable(int64(stacks[i].Score))
+		stacks[i].ViewCountStr = utilities.FormatHumanReadable(int64(stacks[i].ViewCount))
 		stacks[i].CreationDateStr = utilities.FormatDate(stacks[i].CreationDate * 1000)
 	}
